% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{clean_cache}
\alias{clean_cache}
\title{Clean Data Cache}
\usage{
clean_cache(data_provider)
}
\arguments{
\item{data_provider}{An object of type \code{data_provider} (see \code{\link[cake]{datenbaecker}})}
}
\description{
Cleans the cached data in memory and on disk.
}
