% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{datenbaecker}
\alias{datenbaecker}
\alias{local_data_provider}
\alias{remote_data_provider}
\alias{default_data_provider}
\title{Create Data Provider}
\usage{
datenbaecker(cache_dir = NULL)

local_data_provider(cache_dir)

remote_data_provider(host, cache_dir = NULL)

default_data_provider(dp = NULL)
}
\arguments{
\item{cache_dir}{}

\item{dp}{Data provider that should act as default provider. Default is \code{datenbaecker()}.}
}
\description{
Create Data Provider
}
