% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{datenbaecker_theme}
\alias{datenbaecker_theme}
\title{Theme for ggplot2}
\usage{
datenbaecker_theme(...)
}
\arguments{
\item{...}{Arguments for \code{\link[ggplot2]{theme_minimal}}}
}
\value{
The ggplot theme which is also set as default for the session
}
\description{
Theme for ggplot2
}
