% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boundaries.R
\name{geom_swiss_boundaries}
\alias{geom_swiss_boundaries}
\alias{geom_canton}
\alias{geom_commune}
\title{Plot Swiss Boundaries}
\usage{
geom_swiss_boundaries(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  gov_level = c("canton", "commune")
)

geom_canton(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_commune(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}.}

\item{data}{Data with the units for plotting}

\item{data_provider}{A datacake data provider.}

\item{gov_level}{Goverment level (either \dQuote{commune} or \dQuote{canton})}
}
\description{
Plot Swiss Boundaries
}
\section{Aesthetics}{

The only required aesthetics is \code{id} and should contain the
identification for the geographical unit (canton or commune).
If the column is numeric, the values get interpreted as
or canton number according to the definition of the Swiss
Federal Statistical Office (see \url{https://www.bfs.admin.ch/bfs/de/home/grundlagen/agvch/identifikatoren-gemeinde.html}).
Otherwise the column gets interpreted as commune name or
abbreviated canton (ZH, BE, LU, \dots).
The aesthetics \code{x}, \code{y}, \code{group} and \code{subgroup} are ignored.
Otherwise the function supports the same aesthetics as \code{\link[ggplot2]{geom_polygon}}
}

\examples{
\dontrun{
cantons <- tibble(label = c("LU", "ZH"))
ggplot(cantons, aes(id = label)) +
  geom_canton() +
  geom_commune(data = tibble(label = c("Luzern", "Zürich", "Zug")), fill = "white", color = "darkgrey")
}


}
