% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_features.R
\name{get_address_features}
\alias{get_address_features}
\title{Get Analytical Base Table for Swiss Addresses}
\usage{
get_address_features(addresses = NULL, data_provider = default_data_provider())
}
\arguments{
\item{addresses}{Either a character vector with address strings to
search for or \code{NULL} for a standard address sample}

\item{data_provider}{An object of type \code{data_provider}
(see \code{\link[cake]{datenbaecker}})}
}
\value{
A table with a battery of variables for the addresses.
}
\description{
This function provides a set of variables for Swiss addresses.
The data originates from different sources.
You can call \code{\link[cake]{metadata}} on the data frame in order to get
a description and the source of the data in the columns.
If \code{NULL} is passed for the argument \code{addresses}, you get the data
for a selected sample of Swiss addresses.
If address strings are provided, the backend does a fuzzy matching with
the addresses from the
\href{https://www.swisstopo.admin.ch/en/official-directory-of-building-addresses}{official directory of building addresses}
from swisstopo.
}
\examples{
\dontrun{
# This code depends on the availability of the remote server
dp <- datenbaecker(cache_dir = FALSE)
search_for <- c("Bundesstrasse 9 6003 Luzern", "Rämistrasse 71 8006 Zürich")
adr_data <- get_address_features(search_for, data_provider = dp)
}
}
