% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{get_statistical_entities}
\alias{get_statistical_entities}
\alias{get_cantonal_entites}
\alias{get_plz_entites}
\alias{get_communal_entites}
\title{Get Labels and ID for Entities}
\usage{
get_statistical_entities(dp = default_data_provider())

get_cantonal_entites(dp = default_data_provider())

get_plz_entites(dp = default_data_provider())

get_communal_entites(dp = default_data_provider())
}
\arguments{
\item{dp}{An object of type \code{data_provider} (see \code{\link[cake]{datenbaecker}})}
}
\value{
A \code{data.frame} with columns \code{id} and \code{label} containing the
identifiers and labels from the office of federal statistics.
\code{get_statistical_units} returns all entities and has an additional column
\code{entity} to identify the type of entity.
}
\description{
Get Labels and ID for Entities
}
\examples{
\dontrun{
dp <- datenbaecker()
cantons <- get_cantonal(entities)
communes <- get_communal_entities()
}
}
\references{
The source of the data is the swissBOUNDARIES3D dataset from
the Federal Office of Topography swisstopo:
\url{https://www.swisstopo.admin.ch/de/geodata/landscape/boundaries3d.html}
}
