% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{menu_to_cache}
\alias{menu_to_cache}
\title{#' @export
datenbaecker <- function(cache_data = NULL) {
stop("datenbaecker not working yet")
if (is.null(cache_data)) {
cache_data <- ask_cache_data()
}
if (is.logical(cache_data)) {
if (cache_data) {
cache_data_path <- create_default_cache()
} else {
cache_data_path <- NULL
}
} else if (is.character(cache_data)) {
cache_data_path <- normalizePath(cache_data, mustWork = FALSE)
if (!dir.exists(cache_data_path)) {
datacake_abort("invalid_cache_dir", cache_data_path = cache_data_path)
}
} else {
datacake_abort("invalid_cache_data_arg")
}
list(
url = "https://datenbaecker.",
cache_data = cache_data_path
) \%>\%
structure(class = c("datenbaecker", "remote_datacake_provider", "datacake_provider"))
}}
\usage{
menu_to_cache(menu)
}
\description{
#' @export
local_datacake_provider <- function(resource_folder) {
resource_folder <- normalizePath(resource_folder, mustWork = FALSE)
if (!dir.exists(resource_folder)) {
datacake_abort("invalid_cache_dir", cache_data_path = resource_folder)
}
list(
resource_folder = resource_folder
) \%>\%
structure(class = c("local_datacake_provider", "datacake_provider"))
}
}
