% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Get Metadata for an Object}
\usage{
metadata(x, data_provider = default_data_provider(), show = print, ...)
}
\arguments{
\item{x}{Object you want to get metadata for}

\item{data_provider}{An object of type \code{data_provider} (see \code{\link[cake]{datenbaecker}})}

\item{show}{A method to print the metadata (default is \code{print})}

\item{...}{Further arguments that get passed to the \code{show} method}
}
\value{
An invisible data frame with the metadata
}
\description{
Get Metadata for an Object
}
\examples{
\dontrun{
# This code depends on the availability of the remote server
dp <- datenbaecker(cache_dir = FALSE)
adr_feat <- get_address_features(data_provider = dp)
metadata(adr_feat[, c("com_name", "com_canton", "bev_total")])
}
}
