% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_info.R
\name{search_companies}
\alias{search_companies}
\title{Search for Swiss Companies by Name}
\usage{
search_companies(companies, data_provider = default_data_provider())
}
\arguments{
\item{companies}{A character vector with company names}

\item{data_provider}{An object of type \code{data_provider}
(see \code{\link[cake]{datenbaecker}})}
}
\value{
A list with:
\itemize{
\item \code{company_details}: A table with information for the searched companies (legal form, address, etc.)
\item \code{related_companies}: A table with related companies for the search companies
}
}
\description{
Search for Swiss Companies by Name
}
\examples{
\dontrun{
# This code depends on the availability of the remote server
dp <- datenbaecker(cache_dir = FALSE)
search_for <- c("Datenbäcker")
company_data <- search_companies(search_for, data_provider = dp)
}
}
