% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{serve}
\alias{serve}
\title{Get Data from a Data Provider}
\usage{
serve(what, dp, ...)
}
\arguments{
\item{what}{Name of the data source}

\item{dp}{An object of type \code{data_provider} (see \code{\link[datacake]{datenbaecker}})}

\item{...}{Further parameter (currently not used)}
}
\value{
The requested data
}
\description{
Depending on the type of \code{data_provider} and the settings,
\code{serve} might return cached data from memory or disk.
If the data does not exist locally, a \code{remote_data_provider} might download the
data from the internet and saves is to the cache folder on the disk depending
on the settings (see \code{\link[datacake]{datenbaecker}}).
}
