% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{datenbaecker}
\alias{datenbaecker}
\alias{local_data_provider}
\alias{remote_data_provider}
\alias{default_data_provider}
\title{Create Data Provider}
\usage{
datenbaecker(cache_dir = NULL, auth_info = NULL)

local_data_provider(cache_dir)

remote_data_provider(
  host,
  cache_dir = NULL,
  auth_info = NULL,
  api_version_prefix = ""
)

default_data_provider(dp = NULL)
}
\arguments{
\item{cache_dir}{\code{logical} if downloaded data should to be saved to disk or a \code{character} containing a directory path to the cache folder.
The program asks if data should be saved to disk if the argument is \code{NULL}.}

\item{auth_info}{Authentication info for the remote server (currently unused)}

\item{host}{URL to the remote data provider}

\item{api_version_prefix}{Version prefix for the URL}

\item{dp}{Data provider that should act as default provider. Default is \code{datenbaecker()}.}
}
\value{
An object of type \code{data_provider}
}
\description{
Create Data Provider
}
