% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boundaries.R
\name{geom_swiss_boundaries}
\alias{geom_swiss_boundaries}
\alias{geom_canton}
\alias{geom_commune}
\title{Plot Swiss Boundaries}
\usage{
geom_swiss_boundaries(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  gov_level = c("canton", "commune")
)

geom_canton(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_commune(
  mapping = NULL,
  data = NULL,
  data_provider = default_data_provider(),
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}.}

\item{data}{Data with the units for plotting}

\item{data_provider}{A datacake data provider. Default is \code{\link[datacake]{default_data_provider}}.}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{gov_level}{Goverment level (either \dQuote{commune} or \dQuote{canton})}
}
\description{
Plot Swiss Boundaries
}
\section{Aesthetics}{

The only required aesthetics is \code{id} and should contain the
identification for the geographical unit (canton or commune).
If the column is \code{numeric}, the values get interpreted as commune
or canton number according to the definition of the Swiss
Federal Statistical Office (see \url{https://www.bfs.admin.ch/bfs/de/home/grundlagen/agvch/identifikatoren-gemeinde.html}).
Otherwise the column gets interpreted as commune name or
abbreviated canton (ZH, BE, LU, \dots).
The aesthetics \code{x}, \code{y}, \code{group} and \code{subgroup} are ignored.
Otherwise the function supports the same aesthetics as \code{\link[ggplot2]{geom_polygon}}
}

\examples{
\dontrun{
cantons <- tibble(label = c("LU", "ZH"))
ggplot(cantons, aes(id = label)) +
  geom_canton() +
  geom_commune(data = tibble(label = c("Luzern", "Zürich", "Zug")), fill = "white", color = "darkgrey")
}

}
